Moved all serial port related files to separate dir
