+++ biu_ops_subaru_switches.h
@@ -15,7 +15,7 @@
 #include <QGroupBox>
 #include <QLabel>
 
-#include <serial_port_actions.h>
+#include <serial_port/serial_port_actions.h>
 
 QT_BEGIN_NAMESPACE
 namespace Ui
