+++ ecu_operations.h
@@ -9,7 +9,7 @@
 #include <kernelcomms.h>
 #include <kernelmemorymodels.h>
 #include <file_actions.h>
-#include <serial_port_actions.h>
+#include <serial_port/serial_port_actions.h>
 
 class EcuOperations : public QWidget
 {
