+++ ecu_operations_manual.h
@@ -8,7 +8,7 @@
 #include <kernelcomms.h>
 #include <kernelmemorymodels.h>
 #include <file_actions.h>
-#include <serial_port_actions.h>
+#include <serial_port/serial_port_actions.h>
 
 QT_BEGIN_NAMESPACE
 namespace Ui
