+++ modules/flash_ecu_subaru_denso_sh705x_can.h
@@ -14,7 +14,7 @@
 #include <kernelcomms.h>
 #include <kernelmemorymodels.h>
 #include <file_actions.h>
-#include <serial_port_actions.h>
+#include <serial_port/serial_port_actions.h>
 #include <ui_ecu_operations.h>
 
 QT_BEGIN_NAMESPACE
