+++ modules/flash_ecu_subaru_hitachi_m32r_06.h
@@ -13,7 +13,7 @@
 
 #include <kernelmemorymodels.h>
 #include <file_actions.h>
-#include <serial_port_actions.h>
+#include <serial_port/serial_port_actions.h>
 #include <ui_ecu_operations.h>
 
 QT_BEGIN_NAMESPACE
