+++ modules/flash_tcu_subaru_hitachi_m32r_can.h
@@ -16,7 +16,7 @@
 #include <kernelcomms.h>
 #include <kernelmemorymodels.h>
 #include <file_actions.h>
-#include <serial_port_actions.h>
+#include <serial_port/serial_port_actions.h>
 #include <ui_ecu_operations.h>
 
 QT_BEGIN_NAMESPACE
